+++ editor/mokap/core/src/main/java/es/eucm/ead/editor/control/MokapController.java
@@ -41,26 +41,20 @@ import com.badlogic.gdx.Files;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input.Buttons;
 import com.badlogic.gdx.Input.Keys;
-import com.badlogic.gdx.math.Interpolation;
-import com.badlogic.gdx.scenes.scene2d.*;
-import com.badlogic.gdx.scenes.scene2d.actions.Actions;
-import com.badlogic.gdx.scenes.scene2d.ui.Image;
-import com.badlogic.gdx.scenes.scene2d.ui.Skin;
+import com.badlogic.gdx.scenes.scene2d.Group;
+import com.badlogic.gdx.scenes.scene2d.InputEvent;
+import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.ui.TextArea;
-
-import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
-import com.badlogic.gdx.utils.Array;
 import es.eucm.ead.editor.assets.ApplicationAssets;
 import es.eucm.ead.editor.control.engine.MobileEngineInitializer;
 import es.eucm.ead.editor.platform.Platform;
-import es.eucm.ead.editor.view.SkinConstants;
 import es.eucm.ead.editor.view.widgets.SpinningLogo;
 import es.eucm.ead.editor.view.widgets.WidgetBuilder;
 import es.eucm.ead.engine.EngineInitializer;
 
 public class MokapController extends Controller {
 
-	private SpinningLogo loadingImage;
+	private Group loadingImage;
 
 	public static enum Dpi {
 		LDPI(150), MDPI(190), HDPI(270), XHDPI(Float.MAX_VALUE), XXHDPI(0), XXXHDPI(
@@ -149,7 +143,8 @@ public class MokapController extends Controller {
 		});
 
 		loadingImage = new SpinningLogo(getApplicationAssets().getSkin());
-
+		loadingImage.setPosition(WidgetBuilder.dpToPixels(24),
+				WidgetBuilder.dpToPixels(24));
 	}
 
 	public void pause() {
@@ -196,8 +191,6 @@ public class MokapController extends Controller {
 		if (!getViews().modalsContainer.getChildren().contains(loadingImage,
 				true)) {
 			getViews().addToModalsContainer(loadingImage);
-			loadingImage.reset();
-
 		} else {
 			loadingImage.toFront();
 		}
